+++ R/ResamplingHoldout.R
@@ -15,7 +15,7 @@
 #'   Ratio of observations to put into the training set.
 #'
 #' @references
-#' \cite{mlr3}{bischl_2012}
+#' `r tools::toRd(bibentries["bischl_2012"])`
 #'
 #' @template seealso_resampling
 #' @export