+++ R/ResamplingRepeatedCV.R
@@ -22,7 +22,7 @@
 #'   Number of folds.
 #'
 #' @references
-#' \cite{mlr3}{bischl_2012}
+#' `r tools::toRd(bibentries["bischl_2012"])`
 #'
 #' @template seealso_resampling
 #' @export
@@ -118,6 +118,14 @@ ResamplingRepeatedCV = R6Class("ResamplingRepeatedCV", inherit = Resampling,
 
     .combine = function(instances) {
       rbindlist(instances, use.names = TRUE)
+    },
+
+    deep_clone = function(name, value) {
+      switch(name,
+        "instance" = copy(value),
+        "param_set" = value$clone(deep = TRUE),
+        value
+      )
     }
   )
 )