+++ R/as_data_backend.R
@@ -2,16 +2,18 @@
 #'
 #' @description
 #' Wraps a [DataBackend] around data.
+#' \CRANpkg{mlr3} ships with methods for `data.frame` (converted to a [DataBackendDataTable]
+#' and `Matrix` from package \CRANpkg{Matrix} (converted to a [DataBackendMatrix]).
+#'
+#' Additional methods are implemented in the package \CRANpkg{mlr3db}, e.g. to connect
+#' to real DBMS like PostgreSQL (via \CRANpkg{dbplyr}) or DuckDB (via \CRANpkg{DBI}/\CRANpkg{duckdb}).
 #'
 #' @param data `any`\cr
 #'   Data to create a [DataBackend] from.
 #'   For a `data.frame()` (this includes `tibble()` from \CRANpkg{tibble} and [data.table::data.table()]),
 #'   a [DataBackendDataTable] is created.
-#'   See `methods("as_data_backend")` for possible input formats.
-#'
-#'   Package \CRANpkg{mlr3db} extends this function with a method for lazy table objects
-#'   implemented in \CRANpkg{dbplyr}.
-#'   This allows to interface many different data base systems such as SQL servers.
+#'   For objects of type `Matrix` (from package \CRANpkg{Matrix}), a [DataBackendMatrix] is returned.
+#'   See `methods("as_data_backend")` for all possible input formats.
 #'
 #' @template param_primary_key
 #'