+++ R/zzz.R
@@ -10,18 +10,23 @@
 #' @section Additional resources:
 #' * Book on mlr3: \url{https://mlr3book.mlr-org.com}
 #' * Use cases and examples: \url{https://mlr3gallery.mlr-org.com}
+#' * More classification and regression tasks: \CRANpkg{mlr3data}
+#' * Connector to [OpenML](https://www.openml.org): \CRANpkg{mlr3oml}
 #' * More classification and regression learners: \CRANpkg{mlr3learners}
+#' * Even more learners: \url{https://github.com/mlr-org/mlr3extralearners}
 #' * Preprocessing and machine learning pipelines: \CRANpkg{mlr3pipelines}
 #' * Tuning of hyperparameters: \CRANpkg{mlr3tuning}
 #' * Visualizations for many \pkg{mlr3} objects: \CRANpkg{mlr3viz}
 #' * Survival analysis and probabilistic regression: \CRANpkg{mlr3proba}
+#' * Cluster analysis: \CRANpkg{mlr3cluster}
 #' * Feature selection filters: \CRANpkg{mlr3filters}
+#' * Feature selection wrappers: \CRANpkg{mlr3fselect}
 #' * Interface to real (out-of-memory) data bases: \CRANpkg{mlr3db}
 #' * Performance measures as plain functions: \CRANpkg{mlr3measures}
 #' * Parallelization framework: \CRANpkg{future}
 #' * Progress bars: \CRANpkg{progressr}
 #' @references
-#' \cite{mlr3}{pkg::citation}
+#' `r tools::toRd(citation("mlr3"))`
 "_PACKAGE"
 
 dummy_import = function() {