+++ tests/testthat/test_mlr_resamplings.R
@@ -1,8 +1,6 @@
-context("mlr_resamplings")
-
 test_that("mlr_resamplings", {
   expect_dictionary(mlr_resamplings, min_items = 1L)
-  keys = mlr_resamplings$keys()
+  keys = setdiff(mlr_resamplings$keys(), "ordered_holdout")
 
   task = tsk("iris")
   for (key in keys) {