+++ tests/testthat/test_resampling_insample.R
@@ -1,5 +1,3 @@
-context("mlr_resampling_insample")
-
 test_that("insample has no duplicated ids", {
   r = rsmp("insample")
   expect_identical(r$duplicated_ids, FALSE)