+++ NEWS.md
@@ -1,3 +1,11 @@
+# mlr3 0.8.0
+
+* `MeasureOOBError` now has set property `minimize` to `TRUE`.
+* New learner property `"featureless"` to tag learners which can operate on
+  featureless tasks.
+* Fixed [ResampleResult] ignoring argument `predict_sets` for returned
+  [Prediction] objects.
+
 # mlr3 0.7.0
 
 * Updated properties of featureless learners to apply it on all feature types
@@ -9,6 +17,9 @@
 * `ResampleResult` and `BenchmarkResult` now share a common interface to store
   the experiment results. Manual construction is still possible with helper
   function `as_result_data()`
+* Fixed deep cloning of `ResamplingCV` and `ResamplingRepeatedCV`.
+* New measure `classif.prauc` (area under precision-recall curve).
+* Removed dependency on orphaned package `bibtex`.
 
 # mlr3 0.6.0